Rename auth-header.js to authHeader.js - Fix imports
