+++ src/services/auth.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 import { API_URL } from '../constants';
 import storage from '../utils/storage';
-import { authHeader } from './auth-header';
+import { authHeader } from './authHeader';
 
 const baseUrl = `${API_URL}/auth`;
