+++ src/services/tweets.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 import { API_URL } from '../constants';
-import { authHeader } from './auth-header';
+import { authHeader } from './authHeader';
 
 const baseUrl = `${API_URL}/tweets`;
